+++ src/Reactant/Models/Reactant.php
@@ -30,6 +30,7 @@ use Cog\Laravel\Love\Reactant\ReactionTotal\Models\NullReactionTotal;
 use Cog\Laravel\Love\Reactant\ReactionTotal\Models\ReactionTotal;
 use Cog\Laravel\Love\Reaction\Models\Reaction;
 use Cog\Laravel\Love\Support\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\HasOne;
@@ -45,12 +46,13 @@ final class Reactant extends Model implements
 
     protected static $unguarded = true;
 
-    /**
-     * @var string[]
-     */
-    protected $casts = [
-        'id' => 'string',
-    ];
+    public function id(): Attribute
+    {
+        return new Attribute(
+            get: fn (string | null $value) => $value,
+            set: fn (string | null $value) => $value,
+        );
+    }
 
     public function reactable(): MorphTo
     {