+++ src/Reactant/ReactionCounter/Models/ReactionCounter.php
@@ -19,6 +19,7 @@ use Cog\Contracts\Love\ReactionType\Models\ReactionType as ReactionTypeInterface
 use Cog\Laravel\Love\Reactant\Models\Reactant;
 use Cog\Laravel\Love\ReactionType\Models\ReactionType;
 use Cog\Laravel\Love\Support\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
@@ -43,13 +44,21 @@ final class ReactionCounter extends Model implements
         'weight' => self::WEIGHT_DEFAULT,
     ];
 
-    /**
-     * @var string[]
-     */
-    protected $casts = [
-        'count' => 'integer',
-        'weight' => 'float',
-    ];
+    public function count(): Attribute
+    {
+        return new Attribute(
+            get: fn (int | null $value) => $value ?? self::COUNT_DEFAULT,
+            set: fn (int | null $value) => $value ?? self::COUNT_DEFAULT,
+        );
+    }
+
+    public function weight(): Attribute
+    {
+        return new Attribute(
+            get: fn (float | null $value) => $value ?? self::WEIGHT_DEFAULT,
+            set: fn (float | null $value) => $value ?? self::WEIGHT_DEFAULT,
+        );
+    }
 
     public function reactant(): BelongsTo
     {
@@ -115,17 +124,5 @@ final class ReactionCounter extends Model implements
         float $amount,
     ): void {
         $this->decrement('weight', $amount);
-    }
-
-    public function setCountAttribute(
-        int | null $count,
-    ): void {
-        $this->attributes['count'] = $count ?? self::COUNT_DEFAULT;
-    }
-
-    public function setWeightAttribute(
-        float | null $weight,
-    ): void {
-        $this->attributes['weight'] = $weight ?? self::WEIGHT_DEFAULT;
     }
 }